core: Refactor Annotations Processor core: TP_JAVA_FILE
