+++ AnnotationsProcessor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-com.christophecvb.touchportal.annotations.processor.TouchPortalPluginAnnotationProcessor
+com.christophecvb.touchportal.annotations.processor.TouchPortalPluginAnnotationsProcessor
