+++ Library/src/test/resources/TouchPortalPluginTest/plugin.config
@@ -1,4 +1,4 @@
 #TouchPortalPluginTest
-#Mon May 23 10:54:55 CEST 2022
+#Sat Nov 26 23:11:43 CET 2022
 samplekey=Sample Value
 plugin.version=1
